Bug 8744: fix writing ASN.1 IP ID values in GSO IP ID helper
Reading a 16-bit NDN field from its ASN.1 signed 32-bit integer
container to a uint16_t container works correctly. The opposite
doesn't work correctly: one must use int32_t container for this.

Fixes: c47999cc5437 ("Bug 8744: Introduce granular API to cope with Tx offloads")

+++ lib/tapi_tad/tapi_ndn.c
@@ -2347,16 +2347,16 @@ tapi_ndn_gso_pkts_ip_id_edit(asn_value        **pkts,
 
     for (i = 0; i < nb_pkts; ++i)
     {
-        uint16_t provisional_ip_id;
+        int32_t provisional_ip_id;
 
         if (inc_mod15)
         {
-            provisional_ip_id = (superframe_ip_id & 0x8000) |
-                                ((superframe_ip_id + i) & 0x7fff);
+            provisional_ip_id = (uint16_t)((superframe_ip_id & 0x8000) |
+                                           ((superframe_ip_id + i) & 0x7fff));
         }
         else
         {
-            provisional_ip_id = superframe_ip_id + i;
+            provisional_ip_id = (uint16_t)(superframe_ip_id + i);
         }
 
         assert(pkts[i] != NULL);
